backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
httpx==0.27.0
xgboost==2.0.3